# docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.createCompanyRequest:
    properties:
      contact_email:
        example: john@acme.com
        type: string
      contact_name:
        example: John Doe
        type: string
      name:
        example: Acme Inc
        type: string
    required:
    - contact_email
    - contact_name
    - name
    type: object
...
  handlers.joinCompanyRequest:
    properties:
      secret_code:
        example: random hex
        type: string
    required:
    - secret_code
    type: object
...
info:
  contact: {}
  description: This endpoint is for dev purposes
  title: VNEST Dashboard Swagger docs
  version: "1.0"
paths:
...
  /company/create:
    post:
      consumes:
      - application/json
      description: Creates a new company for the user
      parameters:
      - description: Company details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createCompanyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a company
      tags:
      - company
  /company/delete:
    delete:
      description: Deletes the user's company
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a company
      tags:
      - company
  /company/edit:
    put:
      consumes:
      - application/json
      description: Updates the existing company data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Edit company information
      tags:
      - company
  /company/join/{id}:
    post:
      consumes:
      - application/json
      description: Join company for the user
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      - description: Secret Code
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.joinCompanyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Join a company
      tags:
      - company
...
  /company/me:
    get:
      description: Retrieves the company associated with the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user's company
      tags:
      - company
...
  /manage/company/delete/{id}:
    delete:
      description: Deletes a company by its ID (admin only)
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Admin delete company
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" and your JWT token for authentication and authorization
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"